Setting relativeRotation of a shape correctly
When creating a multi-shaped object, the relativeRotation is currently being set as an array. When calculating the mass (in setupMass), the relativeRotation is expected to be a Mat33.
This creates a quaternion from the rotation and sets the Mat33 object from this quaternion. 
It would be cleaner if I added a few help functions (like QuadToMatrix), but since this is not often used, those 2 objects created will not influence performance.