Update MatRotation.js
This is causing some strange results when rotating by low or high or negative numbers on the X Y or Z axis, for example in the Codepen below I have rotated two platforms on the Y axis by 359 and 1 degree so the Red balls should fall and come to a rest on the Blue platforms, instead they roll off in opposite directions.

http://codepen.io/pigloo/pen/ZWaBrR/

This piece of code appears to have been ported from here: http://www.euclideanspace.com/maths/geometry/rotations/conversions/eulerToAngle/ but I think there is an error. 

I am not sure if this effects anything else but from my testing it solves the issue.